# OPENAPI_SWAGGER_SPEC.yaml
openapi: 3.0.0
info:
  title: Asset Management Enterprise SaaS API
  description: Complete API documentation for Asset Management system with enterprise features
  version: 1.0.0
  contact:
    name: API Support
    url: https://example.com/support
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8085/api/v1
    description: Local Development
  - url: https://api.example.com/api/v1
    description: Production
paths:
...
  /dashboard/summary:
    get:
      summary: Get dashboard summary
      operationId: getDashboardSummary
      tags:
        - Dashboard
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'
...
  /departments:
    get:
      summary: List all departments
      operationId: listDepartments
      tags:
          - Departments
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DepartmentDto'
        '401':
          description: Unauthorized
...
  /departments/{id}:
    delete:
      summary: Delete a department
      operationId: deleteDepartment
      tags:
          - Departments
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Department UUID
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
...
  /dashboard/assets-by-status:
    get:
      summary: Asset counts by status
      operationId: getDashboardAssetsByStatus
      tags:
          - Dashboard
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Assets by status
          content:
            application/json:
              schema:
                type: object
                properties:
                  IN_USE:
                    type: integer
                  IN_STORAGE:
                    type: integer
                  DISPOSED:
                    type: integer
                  UNDER_MAINTENANCE:
                    type: integer
        '401':
          description: Unauthorized
  /dashboard/assets-by-department:
    get:
      summary: Asset counts per department
      operationId: getDashboardAssetsByDepartment
      tags:
          - Dashboard
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Assets by department
          content:
            application/json:
              schema:
                type: object
                properties:
                  departments:
                    type: array
        '401':
          description: Unauthorized
  /dashboard/maintenance-alerts:
    get:
      summary: Upcoming and overdue maintenance
      operationId: getDashboardMaintenanceAlerts
      tags:
          - Dashboard
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Maintenance alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  overdue:
                    type: array
                  upcoming:
                    type: array
        '401':
          description: Unauthorized
  /dashboard/depreciation-summary:
    get:
      summary: Depreciation dashboard summary
      operationId: getDashboardDepreciationSummary
      tags:
          - Dashboard
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Depreciation summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalBookValue:
                    type: number
                  totalDepreciation:
                    type: number
                  topDepreciating:
                    type: array
        '401':
          description: Unauthorized
...
  /dpa/consent:
    get:
      summary: List consent records for the organisation
      operationId: listConsents
      tags:
          - DPA
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: Page number
          schema:
            type: integer
        - name: size
          in: query
          required: false
          description: Page size
          schema:
            type: integer
      responses:
        '200':
          description: Paged consents
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                  totalElements:
                    type: integer
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /dpa/consent/check:
    get:
      summary: Check consent for a purpose
      operationId: checkConsent
      tags:
          - DPA
      security:
        - BearerAuth: []
      parameters:
        - name: purpose
          in: query
          required: true
          description: Consent purpose
          schema:
            type: string
      responses:
        '200':
          description: Consent status
          content:
            application/json:
              schema:
                type: object
                properties:
                  granted:
                    type: boolean
                  purpose:
                    type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /dpa/consent/{purpose}:
    delete:
      summary: Revoke consent for a purpose
      operationId: revokeConsent
      tags:
          - DPA
      security:
        - BearerAuth: []
      parameters:
        - name: purpose
          in: path
          required: true
          description: Consent purpose identifier
          schema:
            type: string
      responses:
        '200':
          description: Revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  purpose:
                    type: string
                  granted:
                    type: boolean
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
...
  /dpa/dsar:
    get:
      summary: List DSAR submissions
      operationId: listDsar
      tags:
          - DPA
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          description: Status filter
          schema:
            type: string
        - name: page
          in: query
          required: false
          description: Page number
          schema:
            type: integer
        - name: size
          in: query
          required: false
          description: Page size
          schema:
            type: integer
      responses:
        '200':
          description: Paged DSAR records
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                  totalElements:
                    type: integer
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /dpa/dsar/{id}:
    get:
      summary: Get a DSAR by ID
      operationId: getDsar
      tags:
          - DPA
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: DSAR UUID
          schema:
            type: string
      responses:
        '200':
          description: DSAR
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  requestType:
                    type: string
                  status:
                    type: string
                  submittedAt:
                    type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /dpa/dsar/{id}/status:
    patch:
      summary: Update DSAR processing status
      operationId: updateDsarStatus
      tags:
          - DPA
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: DSAR UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: PENDING|IN_REVIEW|COMPLETED|REJECTED
                notes:
                  type: string
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
...
  /compliance/controls/{id}:
    delete:
      summary: Delete Control
      operationId: deleteCompliance_controls
      tags:
          - Compliance
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Control UUID
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DashboardSummary:
      type: object
      properties:
        totalAssets:
          type: integer
        assetsInUse:
          type: integer
        assetsInStock:
          type: integer
        assetsRetired:
          type: integer
        totalAssetValue:
          type: number
        maintenanceAlerts:
          type: integer
        pendingPurchaseOrders:
          type: integer
...